boxjacobian: inserted function, which writes a new solution back to the global solution vector (localToGlobal) after e.g. a variable switch. This was implemented in the box2p2cjacobian. Moreover, new_newtoncontroller was changed, physicalness is written out correctly (was from previous timestep before)
git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dune-mux/trunk@781 2fb0f335-1f38-0410-981e-8018bf24f1b0

+++ dumux/2p2c/2p2cnewtoncontroller.hh
@@ -97,7 +97,7 @@ namespace Dune {
                 Scalar switchVar;
                 Scalar pW;
 
-                for (int idx = 0; idx < (*u).size(); idx++)
+                for (int idx = 0; idx < (int) (*u).size(); idx++)
                 {
                     pW = (*u)[idx][0];
                 	switchVar = (*u)[idx][1];
